+++ R/modify.R
@@ -16,7 +16,9 @@
 #'
 #' All these functions are S3 generic. However, the default method is
 #' sufficient in many cases. It should be suitable for any data type
-#' that implements the subset-assignment method `[<-`.
+#' that implements the subset-assignment method `[<-`. Methods are provided
+#' for character, logical, integer and double classes (counterparts to `map_chr`,
+#' `map_lgl`, `map_int`, and `map_dbl`)
 #'
 #' In some cases it may make sense to provide a custom implementation
 #' with a method suited to your S3 class. For example, a `grouped_df`
@@ -87,6 +89,26 @@ modify.default <- function(.x, .f, ...) {
   .x[] <- map(.x, .f, ...)
   .x
 }
+#' @export
+modify.integer  <- function (.x, .f, ...) {
+  .x[] <- map_int(.x, .f, ...)
+  .x
+}
+#' @export
+modify.double  <- function (.x, .f, ...) {
+  .x[] <- map_dbl(.x, .f, ...)
+  .x
+}
+#' @export
+modify.character  <- function (.x, .f, ...) {
+  .x[] <- map_chr(.x, .f, ...)
+  .x
+}
+#' @export
+modify.logical  <- function (.x, .f, ...) {
+  .x[] <- map_lgl(.x, .f, ...)
+  .x
+}
 
 #' @export
 modify.pairlist <- function(.x, .f, ...) {
@@ -106,6 +128,30 @@ modify_if.default <- function(.x, .p, .f, ...) {
   .x[sel] <- map(.x[sel], .f, ...)
   .x
 }
+#' @export
+modify_if.integer <- function(.x, .p, .f, ...) {
+  sel <- probe(.x, .p)
+  .x[sel] <- map_int(.x[sel], .f, ...)
+  .x
+}
+#' @export
+modify_if.double <- function(.x, .p, .f, ...) {
+  sel <- probe(.x, .p)
+  .x[sel] <- map_dbl(.x[sel], .f, ...)
+  .x
+}
+#' @export
+modify_if.character <- function(.x, .p, .f, ...) {
+  sel <- probe(.x, .p)
+  .x[sel] <- map_chr(.x[sel], .f, ...)
+  .x
+}
+#' @export
+modify_if.logical <- function(.x, .p, .f, ...) {
+  sel <- probe(.x, .p)
+  .x[sel] <- map_lgl(.x[sel], .f, ...)
+  .x
+}
 
 #' @rdname modify
 #' @export
@@ -117,6 +163,30 @@ modify_at <- function(.x, .at, .f, ...) {
 modify_at.default <- function(.x, .at, .f, ...) {
   sel <- inv_which(.x, .at)
   .x[sel] <- map(.x[sel], .f, ...)
+  .x
+}
+#' @export
+modify_at.integer <- function(.x, .at, .f, ...) {
+  sel <- inv_which(.x, .at)
+  .x[sel] <- map_int(.x[sel], .f, ...)
+  .x
+}
+#' @export
+modify_at.double <- function(.x, .at, .f, ...) {
+  sel <- inv_which(.x, .at)
+  .x[sel] <- map_dbl(.x[sel], .f, ...)
+  .x
+}
+#' @export
+modify_at.character <- function(.x, .at, .f, ...) {
+  sel <- inv_which(.x, .at)
+  .x[sel] <- map_chr(.x[sel], .f, ...)
+  .x
+}
+#' @export
+modify_at.logical <- function(.x, .at, .f, ...) {
+  sel <- inv_which(.x, .at)
+  .x[sel] <- map_lgl(.x[sel], .f, ...)
   .x
 }